Fix in expression and remove between
In Expressions should be set up to do something like this:

    Email.scan.filter(Email.generated_email.in_("test@test.com"))
    ---
    dynamodb scan \
        --table-name "Emails" \
        --filter-expression "generated_email IN (:email1)" \
        --expression-attribute-values '{":email1": {"S": "test@test.com"}}'

A separate class was created for now to handle this logic. Tests were
created, fun was had by all.

I removed between since I am not confident it is correct and want to
refactor a lot of the expression logic to be more robust and hopefully i
wont need an InExpression anymore (or a BetweenExpression down the line)