Remove build() from datatype
Building is somewhat of a legacy concept before mapping and parsing was
designed. It is too confusing to have the datatype be responsible for
building data differently from how it can be parsed.

This also allows for better inheritence. A new method reconstruct() is
called to parse the dict. parsing is now evaluating the data from the
dict.

Ex) reconstruct({'S': 'my string'}) --> parse('my_string')

This creates parity with mapping/resolving <-> reconstructing/parsing

Now all of the responsibility is placed on the parser to interpret the
data for the models. This means initializing a model does not care what
attributes are passed to each field. You could pass a int to a Datetime
field, and a runtime error will get raised when you try to map or
reconstruct it.